requirements.txt
asgiref==3.7.2
Django==4.2.3
django-crispy-forms==1.13.0
django-debug-toolbar==3.7.0
django-environ==0.8.1
django-extensions==3.2.1
django-tables2==2.4.1
Pillow==9.3.0
pydotplus==2.0.2
pyparsing==3.0.9
pytz==2021.3
sqlparse==0.4.4
psycopg2-binary==2.9.7
gunicorn==21.2.0
eventlet==0.33.3
dnspython==2.3.0
djangorestframework==3.14.0
Markdown==3.4.3
django-filter==23.2
django-ipware==5.0.0
python-dotenv==1.0.0

#[api-packages]
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pycodestyle==2.10.0